requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
cytoolz==0.8.2
ethereum-abi-utils==0.4.3
ethereum-keys==0.1.0a7
ethereum-tester==0.1.0b2
ethereum-utils==0.5.1
idna==2.6
py-solc==1.4.0
pylru==1.0.9
pysha3==1.0.2
requests==2.18.4
rlp==0.6.0
semantic-version==2.6.0
toolz==0.8.2
urllib3==1.22
web3==3.16.2
